Repair some game bugs and formating

+++ TheGame/Poker/GameObjects/Player/IPlayer.cs
@@ -14,7 +14,7 @@
 
         string Status { get; set; }
 
-        double PokerHandMultiplier { get; set; }
+        int PokerHandMultiplier { get; set; }
 
         double CardPower { get; set; }
